Extract replacer into own file
